Separated some rendering tasks from main
